feat: setInterval 대신 setTimeout사용 및 모듈화 수정
